generage js file.
