+++ snippets/js/.source
@@ -1,2 +1,2 @@
-TypeScript/src.ts
-JavaScript/src.js
+src.ts
+src.js
