+++ snippets/js/README.md
@@ -1,0 +1,8 @@
+## Requirement
+
+- [TypeScript](https://www.typescriptlang.org/)
+
+- [typings](https://github.com/typings/typings)
+
+- [js-beautify](https://github.com/beautify-web/js-beautify)
+
